# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Adpoint API",
        "contact": {
            "email": "alkhamilnaz@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Laravel Adpoint API"
        },
        {
            "url": "https://127.0.0.1:8000/api",
            "description": "Laravel Adpoint API"
        }
    ],
    "paths": {
        "/brands": {
            "get": {
                "tags": [
                    "Brands"
                ],
                "summary": "Display a listing of the brands",
                "operationId": "list brands",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok"
                    },
                    "401": {
                        "description": "unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/brand/add": {
            "post": {
                "tags": [
                    "Brands"
                ],
                "summary": "Add Brand",
                "operationId": "Add Brand",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BrandSchema"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "ok"
                    },
                    "401": {
                        "description": "unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/brand/{id}": {
            "get": {
                "tags": [
                    "Brands"
                ],
                "summary": "Display a listing of the brand by id",
                "operationId": "list brand by id",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of brand to return",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok"
                    },
                    "401": {
                        "description": "unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Brands"
                ],
                "summary": "Delete brand by id",
                "operationId": "Delete brand by id",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "brand id",
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok"
                    },
                    "401": {
                        "description": "unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/brand/edit/{id}": {
            "put": {
                "tags": [
                    "Brands"
                ],
                "summary": "Edit of the brand by id",
                "operationId": "Edit brand",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of edit brand",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BrandSchema"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "ok"
                    },
                    "401": {
                        "description": "unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "BrandSchema": {
                "title": "Brand",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 0
                    },
                    "name": {
                        "type": "string",
                        "example": "MCM Media"
                    },
                    "logo": {
                        "type": "string",
                        "example": "uploads/logo.jpg"
                    },
                    "slug": {
                        "type": "string",
                        "example": "MCM Media"
                    },
                    "meta_title": {
                        "type": "string",
                        "example": "MCM Media"
                    },
                    "meta_description": {
                        "type": "string",
                        "example": "MCM Media"
                    }
                },
                "type": "object"
            },
...
        },
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}